requirements.txt
setuptools==56.0.0
wheel==0.34.0
numpy==1.20.3
casadi==3.5.5
commonroad-drivability-checker==2022.1
commonroad-io==2022.1
commonroad-rl==2022.2
gym==0.21.0
cvxpy==1.2.1
gurobipy==9.5.2
scipy==1.7.1
omegaconf==2.2.3
protobuf==4.21.5
psutil==5.9.2
pycddlib==2.1.0
pypoman==0.5.4
shapely==1.8.4
stable-baselines3==1.6.0
torch==1.12.1
pyglet==1.5.27 
pyyaml==6.0 
Pillow==8.3.2
sumocr==2022.1
opencv-python
Polygon3==3.0.8
bidict==0.21.2
sparse==0.15.4
xmltodict==0.14.2